moved getFriendlyError to useUiHelpers
