+++ packages/commercetools/theme/composables/useUiHelpers/index.ts
@@ -100,6 +100,29 @@ const useUiHelpers = () => {
 
   const isFacetCheckbox = (): boolean => false;
 
+  const getFriendlyError = (message, email) => {
+
+    const knownErrors = [
+      {
+        originalMessage: `There is already an existing customer with the email '"${email}"'.`,
+        displayMessage: 'There is already an existing customer with this email',
+        fieldName: 'email'
+      },
+      {
+        originalMessage: 'Account with the given credentials not found.',
+        displayMessage: 'Account with the given credentials not found.',
+        fieldName: null
+      },
+      {
+        originalMessage: 'Network error: Customer account with the given credentials not found.',
+        displayMessage: 'Customer account with the given credentials not found.',
+        fieldName: null
+      }
+    ];
+
+    return knownErrors.find(knownError => knownError.originalMessage === message);
+  };
+
   return {
     getFacetsFromURL,
     getCatLink,
@@ -109,7 +132,8 @@ const useUiHelpers = () => {
     setTermForUrl,
     isFacetColor,
     isFacetCheckbox,
-    getSearchTermFromUrl
+    getSearchTermFromUrl,
+    getFriendlyError
   };
 };
