--- packages/commercetools/theme/helpers/errors/index.js
@@ -1,5 +1,0 @@
-import { getFriendlyError } from './authentication';
-
-export {
-  getFriendlyError
-};
